refactor: common asset loading from base url to allow for path
